better integration into the report.js
