+++ lib/commands/report.js
@@ -1,6 +1,5 @@
 const { checkCoverages } = require('./check-coverage')
 const Report = require('../report')
-const monocartReport = require('../monocart-report')
 
 exports.command = 'report'
 
@@ -19,12 +18,6 @@ exports.outputReport = async function (argv) {
     argv.branches = 100
     argv.statements = 100
   }
-
-  if (argv.experimentalMonocart || process.env.EXPERIMENTAL_MONOCART) {
-    await monocartReport(argv)
-    return
-  }
-
   const report = Report({
     include: argv.include,
     exclude: argv.exclude,
@@ -43,7 +36,8 @@ exports.outputReport = async function (argv) {
     src: argv.src,
     skipFull: argv.skipFull,
     excludeNodeModules: argv.excludeNodeModules,
-    mergeAsync: argv.mergeAsync
+    mergeAsync: argv.mergeAsync,
+    monocartArgv: (argv.experimentalMonocart || process.env.EXPERIMENTAL_MONOCART) ? argv : null
   })
   await report.run()
   if (argv.checkCoverage) await checkCoverages(argv, report)
