+++ lib/report.js
@@ -16,6 +16,7 @@ const getSourceMapFromFile = require('./source-map-from-file')
 const v8toIstanbul = require('v8-to-istanbul')
 const util = require('util')
 const debuglog = util.debuglog('c8')
+const { CoverageReport } = require('monocart-coverage-reports')
 
 class Report {
   constructor ({
@@ -36,7 +37,8 @@ class Report {
     allowExternal = false,
     skipFull,
     excludeNodeModules,
-    mergeAsync
+    mergeAsync,
+    monocartArgv
   }) {
     this.reporter = reporter
     this.reporterOptions = reporterOptions || {}
@@ -60,6 +62,7 @@ class Report {
     this.src = this._getSrc(src)
     this.skipFull = skipFull
     this.mergeAsync = mergeAsync
+    this.monocartArgv = monocartArgv
   }
 
   _getSrc (src) {
@@ -73,6 +76,9 @@ class Report {
   }
 
   async run () {
+    if (this.monocartArgv) {
+      return this.runMonocart()
+    }
     const context = libReport.createContext({
       dir: this.reportsDirectory,
       watermarks: this.watermarks,
@@ -87,6 +93,128 @@ class Report {
         ...this.reporterOptions[_reporter]
       }).execute(context)
     }
+  }
+
+  async runMonocart () {
+    const argv = this.monocartArgv
+    const exclude = this.exclude
+
+    function getEntryFilter () {
+      if (argv.entryFilter) {
+        return argv.entryFilter
+      }
+      if (argv.filter) {
+        return argv.filter
+      }
+      return (entry) => {
+        return exclude.shouldInstrument(fileURLToPath(entry.url))
+      }
+    }
+
+    function getSourceFilter () {
+      if (argv.sourceFilter) {
+        return argv.sourceFilter
+      }
+      if (argv.filter) {
+        return argv.filter
+      }
+      return (sourcePath) => {
+        if (argv.excludeAfterRemap) {
+          // console.log(sourcePath)
+          return exclude.shouldInstrument(sourcePath)
+        }
+        return true
+      }
+    }
+
+    function getReports () {
+      const reports = Array.isArray(argv.reporter) ? argv.reporter : [argv.reporter]
+      const reporterOptions = argv.reporterOptions || {}
+
+      return reports.map((reportName) => {
+        const reportOptions = {
+          ...reporterOptions[reportName]
+        }
+        if (reportName === 'text') {
+          reportOptions.skipEmpty = false
+          reportOptions.skipFull = argv.skipFull
+          reportOptions.maxCols = process.stdout.columns || 100
+        }
+        return [reportName, reportOptions]
+      })
+    }
+
+    // --all: add empty coverage for all files
+    function getAllOptions () {
+      if (!argv.all) {
+        return
+      }
+
+      const src = argv.src
+      const workingDirs = Array.isArray(src) ? src : (typeof src === 'string' ? [src] : [process.cwd()])
+      return {
+        dir: workingDirs,
+        filter: (filePath) => {
+          return exclude.shouldInstrument(filePath)
+        }
+      }
+    }
+
+    function initPct (summary) {
+      Object.keys(summary).forEach(k => {
+        if (summary[k].pct === '') {
+          summary[k].pct = 100
+        }
+      })
+      return summary
+    }
+
+    // adapt coverage options
+    const coverageOptions = {
+      logging: argv.logging,
+      name: argv.name,
+      inline: argv.inline,
+      lcov: argv.lcov,
+      outputDir: argv.reportsDir,
+      clean: argv.clean,
+
+      reports: getReports(),
+      all: getAllOptions(),
+
+      // use default value for istanbul
+      defaultSummarizer: 'pkg',
+
+      entryFilter: getEntryFilter(),
+      sourceFilter: getSourceFilter(),
+
+      onEnd: (coverageResults) => {
+        // for check coverage
+        this._allCoverageFiles = {
+          files: () => {
+            return coverageResults.files.map(it => it.sourcePath)
+          },
+          fileCoverageFor: (file) => {
+            const fileCoverage = coverageResults.files.find(it => it.sourcePath === file)
+            return {
+              toSummary: () => {
+                return initPct(fileCoverage.summary)
+              }
+            }
+          },
+          getCoverageSummary: () => {
+            return initPct(coverageResults.summary)
+          }
+        }
+      }
+    }
+    const coverageReport = new CoverageReport(coverageOptions)
+    coverageReport.cleanCache()
+
+    // read v8 coverage data from tempDirectory
+    await coverageReport.addFromDir(argv.tempDirectory)
+
+    // generate report
+    await coverageReport.generate()
   }
 
   async getCoverageMapFromAllCoverageFiles () {
