+++ package.json
@@ -40,7 +40,7 @@
     "istanbul-lib-coverage": "^3.2.0",
     "istanbul-lib-report": "^3.0.1",
     "istanbul-reports": "^3.1.6",
-    "monocart-coverage-reports": "^2.7.10",
+    "monocart-coverage-reports": "^2.8.3",
     "test-exclude": "^6.0.0",
     "v8-to-istanbul": "^9.0.0",
     "yargs": "^17.7.2",
